Openrouteservice Directions

Get directions for different modes of transport.

ors_directions(
  coordinates,
  profile = ors_profile(),
  format = c("geojson", "json", "gpx"),
  ...,
  api_key = ors_api_key(),
  output = c("parsed", "text", "sf")
)

Arguments

coordinates
List of longitude, latitude coordinate pairs visited in order, alternatively a two column matrix or data.frame.

profile
Route profile, defaults to "driving-car".

format
Response format, defaults to "geojson"

...
Optional parameters as described here

api_key
Character scalar containing openrouteservice API key

output
Output format. By default the response is being parsed to a list-based R object

Value

Route between two or more locations in the selected format structured according to output:

for "text", a character vector of length 1 re-encoded to UTF-8.

for "parsed", a parsed R object.

for "sf", a simple features sf object.

Author

Andrzej Oleś <andrzej@openrouteservice.org>

Examples

coordinates = list(c(8.34234, 48.23424), c(8.34423, 48.26424))

# simple call
ors_directions(coordinates, preference="fastest")

#> <ors_directions>
#> List of 4
#>  $ type    : chr "FeatureCollection"
#>  $ features:List of 1
#>   ..$ :List of 4
#>   .. ..$ bbox      : num [1:4] 8.33 48.23 8.35 48.26
#>   .. ..$ type      : chr "Feature"
#>   .. ..$ properties:List of 3
#>   .. .. ..$ segments  :List of 1
#>   .. .. .. ..$ :List of 3
#>   .. .. .. .. ..$ distance: num 5488
#>   .. .. .. .. ..$ duration: num 805
#>   .. .. .. .. ..$ steps   :List of 15
#>   .. .. .. .. .. ..$ :List of 6
#>   .. .. .. .. .. .. ..$ distance   : num 248
#>   .. .. .. .. .. .. ..$ duration   : num 59.4
#>   .. .. .. .. .. .. ..$ type       : int 11
#>   .. .. .. .. .. .. ..$ instruction: chr "Head south on Benatweg"
#>   .. .. .. .. .. .. ..$ name       : chr "Benatweg"
#>   .. .. .. .. .. .. ..$ way_points : int [1:2] 0 9
#>   .. .. .. .. .. ..$ :List of 6
#>   .. .. .. .. .. .. ..$ distance   : num 52.4
#>   .. .. .. .. .. .. ..$ duration   : num 12.6
#>   .. .. .. .. .. .. ..$ type       : int 6
#>   .. .. .. .. .. .. ..$ instruction: chr "Continue straight onto Benatweg"
#>   .. .. .. .. .. .. ..$ name       : chr "Benatweg"
#>   .. .. .. .. .. .. ..$ way_points : int [1:2] 9 12
#>   .. .. .. .. .. ..$ :List of 6
#>   .. .. .. .. .. .. ..$ distance   : num 1689
#>   .. .. .. .. .. .. ..$ duration   : num 135
#>   .. .. .. .. .. .. ..$ type       : int 1
#>   .. .. .. .. .. .. ..$ instruction: chr "Turn right onto Hölzle, K 5528"
#>   .. .. .. .. .. .. ..$ name       : chr "Hölzle, K 5528"
#>   .. .. .. .. .. .. ..$ way_points : int [1:2] 12 62
#>   .. .. .. .. .. ..$ :List of 6
#>   .. .. .. .. .. .. ..$ distance   : num 265
#>   .. .. .. .. .. .. ..$ duration   : num 31.8
#>   .. .. .. .. .. .. ..$ type       : int 13
#>   .. .. .. .. .. .. ..$ instruction: chr "Keep right"
#>   .. .. .. .. .. .. ..$ name       : chr "-"
#>   .. .. .. .. .. .. ..$ way_points : int [1:2] 62 70
#>   .. .. .. .. .. ..$ :List of 6
#>   .. .. .. .. .. .. ..$ distance   : num 303
#>   .. .. .. .. .. .. ..$ duration   : num 36.3
#>   .. .. .. .. .. .. ..$ type       : int 0
#>   .. .. .. .. .. .. ..$ instruction: chr "Turn left"
#>   .. .. .. .. .. .. ..$ name       : chr "-"
#>   .. .. .. .. .. .. ..$ way_points : int [1:2] 70 79
#>   .. .. .. .. .. ..$ :List of 6
#>   .. .. .. .. .. .. ..$ distance   : num 537
#>   .. .. .. .. .. .. ..$ duration   : num 64.4
#>   .. .. .. .. .. .. ..$ type       : int 0
#>   .. .. .. .. .. .. ..$ instruction: chr "Turn left"
#>   .. .. .. .. .. .. ..$ name       : chr "-"
#>   .. .. .. .. .. .. ..$ way_points : int [1:2] 79 91
#>   .. .. .. .. .. .. [list output truncated]
#>   .. .. ..$ summary   :List of 2
#>   .. .. .. ..$ distance: num 5488
#>   .. .. .. ..$ duration: num 805
#>   .. .. ..$ way_points: int [1:2] 0 163
#>   .. ..$ geometry  :List of 2
#>   .. .. ..$ coordinates: num [1:164, 1:2] 8.34 8.34 8.34 8.34 8.34 ...
#>   .. .. ..$ type       : chr "LineString"
#>  $ bbox    : num [1:4] 8.33 48.23 8.35 48.26
#>  $ metadata:List of 6
#>   ..$ attribution   : chr "openrouteservice.org | OpenStreetMap contributors"
#>   ..$ service       : chr "routing"
#>   ..$ timestamp     : num 1.61e+12
#>   ..$ query         :List of 4
#>   .. ..$ coordinates: num [1:2, 1:2] 8.34 8.34 48.23 48.26
#>   .. ..$ profile    : chr "driving-car"
#>   .. ..$ preference : chr "fastest"
#>   .. ..$ format     : chr "geojson"
#>   ..$ engine        :List of 3
#>   .. ..$ version   : chr "6.3.1"
#>   .. ..$ build_date: chr "2020-12-10T15:35:43Z"
#>   .. ..$ graph_date: chr "1970-01-01T00:00:00Z"
#>   ..$ system_message: chr "Preference 'fastest' has been deprecated, using 'recommended'."

# customized options
ors_directions(coordinates, profile="cycling-mountain", elevation=TRUE)

#> <ors_directions>
#> List of 4
#>  $ type    : chr "FeatureCollection"
#>  $ features:List of 1
#>   ..$ :List of 4
#>   .. ..$ bbox      : num [1:6] 8.33 48.23 659.9 8.35 48.26 ...
#>   .. ..$ type      : chr "Feature"
#>   .. ..$ properties:List of 5
#>   .. .. ..$ ascent    : num 125
#>   .. .. ..$ descent   : num 139
#>   .. .. ..$ segments  :List of 1
#>   .. .. .. ..$ :List of 5
#>   .. .. .. .. ..$ distance: num 5215
#>   .. .. .. .. ..$ duration: num 1276
#>   .. .. .. .. ..$ steps   :List of 17
#>   .. .. .. .. .. ..$ :List of 6
#>   .. .. .. .. .. .. ..$ distance   : num 888
#>   .. .. .. .. .. .. ..$ duration   : num 320
#>   .. .. .. .. .. .. ..$ type       : int 11
#>   .. .. .. .. .. .. ..$ instruction: chr "Head southwest"
#>   .. .. .. .. .. .. ..$ name       : chr "-"
#>   .. .. .. .. .. .. ..$ way_points : int [1:2] 0 25
#>   .. .. .. .. .. ..$ :List of 6
#>   .. .. .. .. .. .. ..$ distance   : num 14.3
#>   .. .. .. .. .. .. ..$ duration   : num 2.9
#>   .. .. .. .. .. .. ..$ type       : int 1
#>   .. .. .. .. .. .. ..$ instruction: chr "Turn right onto Sulzbacher Straße, K 5528"
#>   .. .. .. .. .. .. ..$ name       : chr "Sulzbacher Straße, K 5528"
#>   .. .. .. .. .. .. ..$ way_points : int [1:2] 25 26
#>   .. .. .. .. .. ..$ :List of 6
#>   .. .. .. .. .. .. ..$ distance   : num 321
#>   .. .. .. .. .. .. ..$ duration   : num 64.2
#>   .. .. .. .. .. .. ..$ type       : int 6
#>   .. .. .. .. .. .. ..$ instruction: chr "Continue straight onto Sulzbacher Straße, K 5528"
#>   .. .. .. .. .. .. ..$ name       : chr "Sulzbacher Straße, K 5528"
#>   .. .. .. .. .. .. ..$ way_points : int [1:2] 26 35
#>   .. .. .. .. .. ..$ :List of 6
#>   .. .. .. .. .. .. ..$ distance   : num 493
#>   .. .. .. .. .. .. ..$ duration   : num 98.5
#>   .. .. .. .. .. .. ..$ type       : int 4
#>   .. .. .. .. .. .. ..$ instruction: chr "Turn slight left onto Sulzbacher Straße, K 5528"
#>   .. .. .. .. .. .. ..$ name       : chr "Sulzbacher Straße, K 5528"
#>   .. .. .. .. .. .. ..$ way_points : int [1:2] 35 50
#>   .. .. .. .. .. ..$ :List of 6
#>   .. .. .. .. .. .. ..$ distance   : num 265
#>   .. .. .. .. .. .. ..$ duration   : num 59.7
#>   .. .. .. .. .. .. ..$ type       : int 13
#>   .. .. .. .. .. .. ..$ instruction: chr "Keep right"
#>   .. .. .. .. .. .. ..$ name       : chr "-"
#>   .. .. .. .. .. .. ..$ way_points : int [1:2] 50 58
#>   .. .. .. .. .. ..$ :List of 6
#>   .. .. .. .. .. .. ..$ distance   : num 303
#>   .. .. .. .. .. .. ..$ duration   : num 68.1
#>   .. .. .. .. .. .. ..$ type       : int 0
#>   .. .. .. .. .. .. ..$ instruction: chr "Turn left"
#>   .. .. .. .. .. .. ..$ name       : chr "-"
#>   .. .. .. .. .. .. ..$ way_points : int [1:2] 58 67
#>   .. .. .. .. .. .. [list output truncated]
#>   .. .. .. .. ..$ ascent  : num 125
#>   .. .. .. .. ..$ descent : num 139
#>   .. .. ..$ summary   :List of 2
#>   .. .. .. ..$ distance: num 5215
#>   .. .. .. ..$ duration: num 1276
#>   .. .. ..$ way_points: int [1:2] 0 151
#>   .. ..$ geometry  :List of 2
#>   .. .. ..$ coordinates: num [1:152, 1:3] 8.34 8.34 8.34 8.34 8.34 ...
#>   .. .. ..$ type       : chr "LineString"
#>  $ bbox    : num [1:6] 8.33 48.23 659.9 8.35 48.26 ...
#>  $ metadata:List of 5
#>   ..$ attribution: chr "openrouteservice.org | OpenStreetMap contributors"
#>   ..$ service    : chr "routing"
#>   ..$ timestamp  : num 1.61e+12
#>   ..$ query      :List of 4
#>   .. ..$ coordinates: num [1:2, 1:2] 8.34 8.34 48.23 48.26
#>   .. ..$ profile    : chr "cycling-mountain"
#>   .. ..$ format     : chr "geojson"
#>   .. ..$ elevation  : logi TRUE
#>   ..$ engine     :List of 3
#>   .. ..$ version   : chr "6.3.0"
#>   .. ..$ build_date: chr "2020-12-10T13:13:58Z"
#>   .. ..$ graph_date: chr "1970-01-01T00:00:00Z"

# list of locations as `data.frame` output as simple features `sf` object
locations <- data.frame(lng = c(8.34234, 8.327807, 8.34423),
                        lat = c(48.23424, 48.239368, 48.26424))
ors_directions(locations, output = "sf")

#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                          segments
#> 1 2031.6, 327.4, 247.5, 59.4, 11, Head south on Benatweg, Benatweg, 0, 9, 52.4, 12.6, 6, Continue straight onto Benatweg, Benatweg, 9, 12, 824, 65.9, 1, Turn right onto Hölzle, K 5528, Hölzle, K 5528, 12, 35, 790.9, 169.5, 12, Keep left, -, 35, 71, 49.1, 11.8, 4, Turn slight left, -, 71, 72, 67.7, 8.1, 1, Turn right, -, 72, 75, 0, 0, 10, Arrive at your destination, on the right, -, 75, 75, 4264.8, 701.3, 171.1, 20.5, 11, Head northwest, -, 75, 81, 449.8, 54, 13, Keep right, -, 81, 87, 312.9, 46.5, 0, Turn left, -, 87, 98, 6.8, 1.6, 5, Turn slight right onto Sulzbacher Straße, Sulzbacher Straße, 98, 99, 48.8, 7.6, 1, Turn right onto Sulzbacher Straße, K 5528, Sulzbacher Straße, K 5528, 99, 103, 444.5, 53.3, 0, Turn left onto Sulzbacher Straße, Sulzbacher Straße, 103, 116, 436.6, 52.4, 0, Turn left, -, 116, 126, 247.6, 29.7, 13, Keep right, -, 126, 135, 293.7, 52.9, 1, Turn right, -, 135, 145, 252.4, 45.4, 12, Keep left, -, 145, 151, 370.6, 67.6, 13, Keep right, -, 151, 160, 572.1, 111.9, 12, Keep left, -, 160, 176, 37.6, 9, 0, Turn left, -, 176, 179, 469.9, 112.8, 12, Keep left, -, 179, 193, 150.6, 36.1, 2, Turn sharp left, -, 193, 198, 0, 0, 10, Arrive at your destination, on the right, -, 198, 198
#>          summary way_points
#> 1 6296.4, 1028.7 0, 75, 198
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                            geometry
#> 1 8.344268, 8.344147, 8.344098, 8.343945, 8.343853, 8.343320, 8.343098, 8.342861, 8.342711, 8.342328, 8.342045, 8.341843, 8.341711, 8.341092, 8.340386, 8.340280, 8.339983, 8.339451, 8.339300, 8.339219, 8.339185, 8.339372, 8.339367, 8.339262, 8.338886, 8.338431, 8.338327, 8.338225, 8.338175, 8.337919, 8.337250, 8.336922, 8.336726, 8.336421, 8.336210, 8.336115, 8.335913, 8.335782, 8.335572, 8.335367, 8.335175, 8.335038, 8.334949, 8.334932, 8.334887, 8.334801, 8.334613, 8.334327, 8.334110, 8.333989, 8.333893, 8.333611, 8.333423, 8.333357, 8.333309, 8.333291, 8.333338, 8.333455, 8.333424, 8.333324, 8.333182, 8.332941, 8.332787, 8.332689, 8.332346, 8.332037, 8.331659, 8.330940, 8.330617, 8.330020, 8.329769, 8.329540, 8.328893, 8.328359, 8.328209, 8.328164, 8.327920, 8.327629, 8.327353, 8.326981, 8.326494, 8.326338, 8.326237, 8.326189, 8.326209, 8.326291, 8.327031, 8.329405, 8.329363, 8.329271, 8.328481, 8.328122, 8.327613, 8.327364, 8.327197, 8.327086, 8.326883, 8.326803, 8.326608, 8.326629, 8.326886, 8.327037, 8.327127, 8.327241, 8.327144, 8.327138, 8.327258, 8.327458, 8.327598, 8.328226, 8.328795, 8.329717, 8.329955, 8.330263, 8.330731, 8.330975, 8.331258, 8.330556, 8.329865, 8.329128, 8.328832, 8.328606, 8.328412, 8.328342, 8.328084, 8.327975, 8.327820, 8.327715, 8.327707, 8.327735, 8.327871, 8.328022, 8.328051, 8.328067, 8.328094, 8.328188, 8.328516, 8.329388, 8.329619, 8.329930, 8.330171, 8.330683, 8.330780, 8.330909, 8.331049, 8.331165, 8.331417, 8.331592, 8.331745, 8.331778, 8.331733, 8.331604, 8.332141, 8.332452, 8.332688, 8.332668, 8.332765, 8.332860, 8.333589, 8.333881, 8.334074, 8.334615, 8.334832, 8.335546, 8.335655, 8.335753, 8.335783, 8.336230, 8.337095, 8.337525, 8.337830, 8.337898, 8.337994, 8.338356, 8.338735, 8.339091, 8.339439, 8.339604, 8.339748, 8.339754, 8.339960, 8.340984, 8.341287, 8.341604, 8.342296, 8.343026, 8.343188, 8.343387, 8.343800, 8.343999, 8.344228, 8.344626, 8.344987, 8.345244, 8.344865, 8.344067, 8.343897, 8.343478, 8.343433, 48.233826, 48.233507, 48.233435, 48.233136, 48.233047, 48.232736, 48.232473, 48.232307, 48.232240, 48.232159, 48.232090, 48.232035, 48.231946, 48.232163, 48.232388, 48.232450, 48.232740, 48.233150, 48.233316, 48.233457, 48.233646, 48.234329, 48.234539, 48.234685, 48.234971, 48.235181, 48.235280, 48.235773, 48.235892, 48.236102, 48.236483, 48.236771, 48.237039, 48.237391, 48.237641, 48.237759, 48.237947, 48.238040, 48.238146, 48.238292, 48.238458, 48.238638, 48.238881, 48.238927, 48.239159, 48.239268, 48.239367, 48.239550, 48.239787, 48.239891, 48.239936, 48.239899, 48.239852, 48.239836, 48.239847, 48.239893, 48.240178, 48.240437, 48.240587, 48.240681, 48.240740, 48.240727, 48.240677, 48.240618, 48.240159, 48.239983, 48.239855, 48.239675, 48.239559, 48.239262, 48.239232, 48.239282, 48.239183, 48.239424, 48.239499, 48.239540, 48.239765, 48.239960, 48.240086, 48.240206, 48.240318, 48.240411, 48.240579, 48.240698, 48.240779, 48.240899, 48.241608, 48.243751, 48.243836, 48.243930, 48.244383, 48.244589, 48.244880, 48.245066, 48.245277, 48.245423, 48.245555, 48.245606, 48.245810, 48.245870, 48.245936, 48.245996, 48.246012, 48.246020, 48.246266, 48.246502, 48.246696, 48.246832, 48.246865, 48.247012, 48.247227, 48.247518, 48.247633, 48.247912, 48.248155, 48.248308, 48.248526, 48.248909, 48.249228, 48.249545, 48.249737, 48.249949, 48.250198, 48.250322, 48.250757, 48.251030, 48.251499, 48.251941, 48.252051, 48.252168, 48.252433, 48.252827, 48.252982, 48.253367, 48.253482, 48.253678, 48.253748, 48.253956, 48.254050, 48.254114, 48.254189, 48.254523, 48.254627, 48.254812, 48.255072, 48.255189, 48.255350, 48.255536, 48.255884, 48.256163, 48.256781, 48.257332, 48.257903, 48.258317, 48.258781, 48.259148, 48.259448, 48.259582, 48.259789, 48.259898, 48.259932, 48.260141, 48.260261, 48.260712, 48.260829, 48.260994, 48.261319, 48.261624, 48.261891, 48.262004, 48.262411, 48.262441, 48.262433, 48.262320, 48.262012, 48.261771, 48.261581, 48.261778, 48.261829, 48.261830, 48.262052, 48.262661, 48.262828, 48.262945, 48.263073, 48.263176, 48.263176, 48.263132, 48.262989, 48.262970, 48.263014, 48.263142, 48.263166, 48.263242, 48.263233, 48.263207, 48.263233, 48.263529, 48.263552