Openrouteservice Matrix

Obtain one-to-many, many-to-one and many-to-many matrices for time and distance.

ors_matrix(
  locations,
  profile = ors_profile(),
  ...,
  api_key = ors_api_key(),
  output = c("parsed", "text")
)

Arguments

locations
List of longitude, latitude coordinate pairs, alternatively a two column matrix or data.frame.

profile
Route profile, defaults to "driving-car".

...
Optional parameters as described here

api_key
Character scalar containing openrouteservice API key

output
Output format. By default the response is being parsed to a list-based R object

Value

Duration or distance matrix for multiple source and destination

for "text", a character vector of length 1 re-encoded to UTF-8.

for "parsed", a parsed R object.

Author

Andrzej Oleś andrzej@openrouteservice.org

Examples

coordinates = list(
  c(9.970093, 48.477473),
  c(9.207916, 49.153868),
  c(37.573242, 55.801281),
  c(115.663757,38.106467)
)

# query for duration and distance in km
res = ors_matrix(coordinates, metrics = c("duration", "distance"), units = "km")

# duration in hours
res$durations / 3600

#>            [,1]      [,2]     [,3]      [,4]
#> [1,]   0.000000   1.58440 24.21797 109.52353
#> [2,]   1.603047   0.00000 24.03252 109.33808
#> [3,]  24.169111  23.96979  0.00000  86.28591
#> [4,] 109.859183 109.65986 86.54789   0.00000

# distance in km
res$distances

#>          [,1]     [,2]    [,3]     [,4]
#> [1,]     0.00   153.51 2411.25 10076.28
#> [2,]   153.83     0.00 2385.67 10050.69
#> [3,]  2357.82  2333.65    0.00  7688.75
#> [4,] 10045.50 10021.33 7708.21     0.00