Openrouteservice Directions
Source: R/directions.R

Get directions for different modes of transport.

ors_directions(coordinates, profile = c("driving-car", "driving-hgv",
  "cycling-regular", "cycling-road", "cycling-safe", "cycling-mountain",
  "cycling-tour", "cycling-electric", "foot-walking", "foot-hiking",
  "wheelchair"), ...)

Arguments

coordinates
List of longitude, latitude coordinate pairs visited in order.

profile
Route profile, defaults to driving-car.

...
Optional parameters as described here.

Value

Route between two or more locations for a selected profile and its settings as GeoJSON response.

Examples

coordinates = list(c(8.34234, 48.23424), c(8.34423, 48.26424))

# simple call
ors_directions(coordinates, preference="fastest")

#> <ors_directions>
#> List of 3
#>  $ routes:List of 1
#>   ..$ :List of 6
#>   .. ..$ summary        :List of 2
#>   .. .. ..$ distance: num 5800
#>   .. .. ..$ duration: num 857
#>   .. ..$ geometry_format: chr "encodedpolyline"
#>   .. ..$ geometry       : chr "mtkeHuv|q@~@VLHz@\\PR|@hBt@j@^n@L\\NjALv@Jh@NXi@zBm@jCKTy@z@qAhBa@\\[Ne@DgCc@i@?[Ty@hAi@zASRi@R}@H_@N[b@kAdCy@`"| __truncated__
#>   .. ..$ segments       :List of 1
#>   .. .. ..$ :List of 3
#>   .. .. .. ..$ distance: num 5800
#>   .. .. .. ..$ duration: num 857
#>   .. .. .. ..$ steps   :List of 14
#>   .. .. .. .. ..$ :List of 6
#>   .. .. .. .. .. ..$ distance   : int 255
#>   .. .. .. .. .. ..$ duration   : num 61.2
#>   .. .. .. .. .. ..$ type       : int 11
#>   .. .. .. .. .. ..$ instruction: chr "Head south on Benatweg"
#>   .. .. .. .. .. ..$ name       : chr "Benatweg"
#>   .. .. .. .. .. ..$ way_points : int [1:2] 0 9
#>   .. .. .. .. ..$ :List of 6
#>   .. .. .. .. .. ..$ distance   : num 52.4
#>   .. .. .. .. .. ..$ duration   : num 12.6
#>   .. .. .. .. .. ..$ type       : int 6
#>   .. .. .. .. .. ..$ instruction: chr "Continue straight onto Benatweg"
#>   .. .. .. .. .. ..$ name       : chr "Benatweg"
#>   .. .. .. .. .. ..$ way_points : int [1:2] 9 12
#>   .. .. .. .. ..$ :List of 6
#>   .. .. .. .. .. ..$ distance   : num 731
#>   .. .. .. .. .. ..$ duration   : num 52.7
#>   .. .. .. .. .. ..$ type       : int 1
#>   .. .. .. .. .. ..$ instruction: chr "Turn right onto Hölzle, K 5528"
#>   .. .. .. .. .. ..$ name       : chr "Hölzle, K 5528"
#>   .. .. .. .. .. ..$ way_points : int [1:2] 12 33
#>   .. .. .. .. ..$ :List of 6
#>   .. .. .. .. .. ..$ distance   : num 1324
#>   .. .. .. .. .. ..$ duration   : num 95.3
#>   .. .. .. .. .. ..$ type       : int 6
#>   .. .. .. .. .. ..$ instruction: chr "Continue straight onto Sulzbacher Straße, K 5528"
#>   .. .. .. .. .. ..$ name       : chr "Sulzbacher Straße, K 5528"
#>   .. .. .. .. .. ..$ way_points : int [1:2] 33 71
#>   .. .. .. .. ..$ :List of 6
#>   .. .. .. .. .. ..$ distance   : num 444
#>   .. .. .. .. .. ..$ duration   : num 53.3
#>   .. .. .. .. .. ..$ type       : int 1
#>   .. .. .. .. .. ..$ instruction: chr "Turn right onto Sulzbacher Straße"
#>   .. .. .. .. .. ..$ name       : chr "Sulzbacher Straße"
#>   .. .. .. .. .. ..$ way_points : int [1:2] 71 84
#>   .. .. .. .. ..$ :List of 6
#>   .. .. .. .. .. ..$ distance   : num 437
#>   .. .. .. .. .. ..$ duration   : num 52.4
#>   .. .. .. .. .. ..$ type       : int 0
#>   .. .. .. .. .. ..$ instruction: chr "Turn left"
#>   .. .. .. .. .. ..$ name       : chr ""
#>   .. .. .. .. .. ..$ way_points : int [1:2] 84 94
#>   .. .. .. .. .. [list output truncated]
#>   .. ..$ way_points     : int [1:2] 0 183
#>   .. ..$ bbox           : num [1:4] 8.33 48.23 8.35 48.26
#>  $ bbox  : num [1:4] 8.33 48.23 8.35 48.26
#>  $ info  :List of 6
#>   ..$ attribution      : chr "openrouteservice.org | OpenStreetMap contributors"
#>   ..$ osm_file_md5_hash: chr "2334169eacb30828c11eaa79f6ec4af1"
#>   ..$ engine           :List of 2
#>   .. ..$ version   : chr "4.5.0"
#>   .. ..$ build_date: chr "2018-03-26T14:15:39Z"
#>   ..$ service          : chr "routing"
#>   ..$ timestamp        : num 1.52e+12
#>   ..$ query            :List of 11
#>   .. ..$ profile            : chr "driving-car"
#>   .. ..$ preference         : chr "fastest"
#>   .. ..$ coordinates        : num [1:2, 1:2] 8.34 8.34 48.23 48.26
#>   .. ..$ language           : chr "en"
#>   .. ..$ units              : chr "m"
#>   .. ..$ geometry           : logi TRUE
#>   .. .. [list output truncated]

#customized options
ors_directions(coordinates, profile="cycling-mountain", elevation=TRUE, format="geojson")

#> <ors_directions>
#> List of 4
#>  $ features:List of 1
#>   ..$ :List of 4
#>   .. ..$ geometry  :List of 2
#>   .. .. ..$ coordinates: num [1:213, 1:3] 8.34 8.34 8.34 8.34 8.34 ...
#>   .. .. ..$ type       : chr "LineString"
#>   .. ..$ id        : chr "fid-3e73a606_162cc0bb862_-7f25"
#>   .. ..$ type      : chr "Feature"
#>   .. ..$ properties:List of 4
#>   .. .. ..$ summary   :List of 1
#>   .. .. .. ..$ :List of 4
#>   .. .. .. .. ..$ distance: int 6848
#>   .. .. .. .. ..$ duration: num 2129
#>   .. .. .. .. ..$ ascent  : num 230
#>   .. .. .. .. ..$ descent : int 277
#>   .. .. ..$ bbox      : num [1:4] 8.33 48.23 8.35 48.26
#>   .. .. ..$ way_points: int [1:2] 0 212
#>   .. .. ..$ segments  :List of 1
#>   .. .. .. ..$ :List of 5
#>   .. .. .. .. ..$ distance: int 6848
#>   .. .. .. .. ..$ duration: num 2129
#>   .. .. .. .. ..$ ascent  : num 230
#>   .. .. .. .. ..$ descent : int 277
#>   .. .. .. .. ..$ steps   :List of 20
#>   .. .. .. .. .. ..$ :List of 6
#>   .. .. .. .. .. .. ..$ distance   : int 110
#>   .. .. .. .. .. .. ..$ duration   : int 22
#>   .. .. .. .. .. .. ..$ type       : int 11
#>   .. .. .. .. .. .. ..$ instruction: chr "Head east on Benatweg"
#>   .. .. .. .. .. .. ..$ name       : chr "Benatweg"
#>   .. .. .. .. .. .. ..$ way_points : int [1:2] 0 7
#>   .. .. .. .. .. ..$ :List of 6
#>   .. .. .. .. .. .. ..$ distance   : num 56.7
#>   .. .. .. .. .. .. ..$ duration   : num 11.3
#>   .. .. .. .. .. .. ..$ type       : int 5
#>   .. .. .. .. .. .. ..$ instruction: chr "Turn slight right onto Benatweg"
#>   .. .. .. .. .. .. ..$ name       : chr "Benatweg"
#>   .. .. .. .. .. .. ..$ way_points : int [1:2] 7 10
#>   .. .. .. .. .. ..$ :List of 6
#>   .. .. .. .. .. .. ..$ distance   : num 458
#>   .. .. .. .. .. .. ..$ duration   : num 285
#>   .. .. .. .. .. .. ..$ type       : int 4
#>   .. .. .. .. .. .. ..$ instruction: chr "Continue slight left onto Imbrand"
#>   .. .. .. .. .. .. ..$ name       : chr "Imbrand"
#>   .. .. .. .. .. .. ..$ way_points : int [1:2] 10 36
#>   .. .. .. .. .. ..$ :List of 6
#>   .. .. .. .. .. .. ..$ distance   : num 437
#>   .. .. .. .. .. .. ..$ duration   : num 394
#>   .. .. .. .. .. .. ..$ type       : int 2
#>   .. .. .. .. .. .. ..$ instruction: chr "Turn sharp left onto Imbrand"
#>   .. .. .. .. .. .. ..$ name       : chr "Imbrand"
#>   .. .. .. .. .. .. ..$ way_points : int [1:2] 36 50
#>   .. .. .. .. .. ..$ :List of 6
#>   .. .. .. .. .. .. ..$ distance   : num 245
#>   .. .. .. .. .. .. ..$ duration   : num 158
#>   .. .. .. .. .. .. ..$ type       : int 6
#>   .. .. .. .. .. .. ..$ instruction: chr "Continue straight onto Imbrand"
#>   .. .. .. .. .. .. ..$ name       : chr "Imbrand"
#>   .. .. .. .. .. .. ..$ way_points : int [1:2] 50 57
#>   .. .. .. .. .. ..$ :List of 6
#>   .. .. .. .. .. .. ..$ distance   : num 358
#>   .. .. .. .. .. .. ..$ duration   : num 37.9
#>   .. .. .. .. .. .. ..$ type       : int 5
#>   .. .. .. .. .. .. ..$ instruction: chr "Turn slight right onto Hölzleshof"
#>   .. .. .. .. .. .. ..$ name       : chr "Hölzleshof"
#>   .. .. .. .. .. .. ..$ way_points : int [1:2] 57 65
#>   .. .. .. .. .. .. [list output truncated]
#>  $ bbox    : num [1:4] 8.33 48.23 8.35 48.26
#>  $ type    : chr "FeatureCollection"
#>  $ info    :List of 6
#>   ..$ attribution      : chr "openrouteservice.org | OpenStreetMap contributors"
#>   ..$ osm_file_md5_hash: chr "2334169eacb30828c11eaa79f6ec4af1"
#>   ..$ engine           :List of 2
#>   .. ..$ version   : chr "4.5.0"
#>   .. ..$ build_date: chr "2018-03-26T12:51:50Z"
#>   ..$ service          : chr "routing"
#>   ..$ timestamp        : num 1.52e+12
#>   ..$ query            :List of 11
#>   .. ..$ profile            : chr "cycling-mountain"
#>   .. ..$ preference         : chr "fastest"
#>   .. ..$ coordinates        : num [1:2, 1:3] 8.34 8.34 48.23 48.26 0 ...
#>   .. ..$ language           : chr "en"
#>   .. ..$ units              : chr "m"
#>   .. ..$ geometry           : logi TRUE
#>   .. .. [list output truncated]