Openrouteservice Directions

Get directions for different modes of transport.

ors_directions(coordinates, profile = ors_profile(),
  format = c("geojson", "json", "gpx"), ..., api_key = ors_api_key(),
  output = c("parsed", "text", "sf"))

Arguments

coordinates
List of longitude, latitude coordinate pairs visited in order, alternatively a two column matrix or data.frame.

profile
Route profile, defaults to "driving-car".

format
Response format, defaults to "geojson"

...
Optional parameters as described here

api_key
Character scalar containing openrouteservice API key

output
Output format. By default the response is being parsed to a list-based R object

Value

Route between two or more locations in the selected format structured according to output:

for "text", a character vector of length 1 re-encoded to UTF-8.

for "parsed", a parsed R object.

for "sf", a simple features sf object.

Examples

coordinates = list(c(8.34234, 48.23424), c(8.34423, 48.26424))

# simple call
ors_directions(coordinates, preference="fastest")

#> <ors_directions>
#> List of 4
#>  $ type    : chr "FeatureCollection"
#>  $ features:List of 1
#>   ..$ :List of 4
#>   .. ..$ bbox      : num [1:4] 8.33 48.23 8.35 48.26
#>   .. ..$ type      : chr "Feature"
#>   .. ..$ properties:List of 3
#>   .. .. ..$ segments  :List of 1
#>   .. .. .. ..$ :List of 3
#>   .. .. .. .. ..$ distance: num 5510
#>   .. .. .. .. ..$ duration: num 811
#>   .. .. .. .. ..$ steps   :List of 15
#>   .. .. .. .. .. ..$ :List of 6
#>   .. .. .. .. .. .. ..$ distance   : num 255
#>   .. .. .. .. .. .. ..$ duration   : num 61.2
#>   .. .. .. .. .. .. ..$ type       : int 11
#>   .. .. .. .. .. .. ..$ instruction: chr "Head south on Benatweg"
#>   .. .. .. .. .. .. ..$ name       : chr "Benatweg"
#>   .. .. .. .. .. .. ..$ way_points : int [1:2] 0 9
#>   .. .. .. .. .. ..$ :List of 6
#>   .. .. .. .. .. .. ..$ distance   : num 52.4
#>   .. .. .. .. .. .. ..$ duration   : num 12.6
#>   .. .. .. .. .. .. ..$ type       : int 6
#>   .. .. .. .. .. .. ..$ instruction: chr "Continue straight onto Benatweg"
#>   .. .. .. .. .. .. ..$ name       : chr "Benatweg"
#>   .. .. .. .. .. .. ..$ way_points : int [1:2] 9 12
#>   .. .. .. .. .. ..$ :List of 6
#>   .. .. .. .. .. .. ..$ distance   : num 1689
#>   .. .. .. .. .. .. ..$ duration   : num 135
#>   .. .. .. .. .. .. ..$ type       : int 1
#>   .. .. .. .. .. .. ..$ instruction: chr "Turn right onto Hölzle, K 5528"
#>   .. .. .. .. .. .. ..$ name       : chr "Hölzle, K 5528"
#>   .. .. .. .. .. .. ..$ way_points : int [1:2] 12 63
#>   .. .. .. .. .. ..$ :List of 6
#>   .. .. .. .. .. .. ..$ distance   : num 265
#>   .. .. .. .. .. .. ..$ duration   : num 31.8
#>   .. .. .. .. .. .. ..$ type       : int 13
#>   .. .. .. .. .. .. ..$ instruction: chr "Keep right"
#>   .. .. .. .. .. .. ..$ name       : chr "-"
#>   .. .. .. .. .. .. ..$ way_points : int [1:2] 63 71
#>   .. .. .. .. .. ..$ :List of 6
#>   .. .. .. .. .. .. ..$ distance   : num 303
#>   .. .. .. .. .. .. ..$ duration   : num 36.3
#>   .. .. .. .. .. .. ..$ type       : int 0
#>   .. .. .. .. .. .. ..$ instruction: chr "Turn left"
#>   .. .. .. .. .. .. ..$ name       : chr "-"
#>   .. .. .. .. .. .. ..$ way_points : int [1:2] 71 80
#>   .. .. .. .. .. ..$ :List of 6
#>   .. .. .. .. .. .. ..$ distance   : num 537
#>   .. .. .. .. .. .. ..$ duration   : num 64.4
#>   .. .. .. .. .. .. ..$ type       : int 0
#>   .. .. .. .. .. .. ..$ instruction: chr "Turn left"
#>   .. .. .. .. .. .. ..$ name       : chr "-"
#>   .. .. .. .. .. .. ..$ way_points : int [1:2] 80 92
#>   .. .. .. .. .. .. [list output truncated]
#>   .. .. ..$ summary   :List of 2
#>   .. .. .. ..$ distance: num 5510
#>   .. .. .. ..$ duration: num 811
#>   .. .. ..$ way_points: int [1:2] 0 164
#>   .. ..$ geometry  :List of 2
#>   .. .. ..$ coordinates: num [1:165, 1:2] 8.34 8.34 8.34 8.34 8.34 ...
#>   .. .. ..$ type       : chr "LineString"
#>  $ bbox    : num [1:4] 8.33 48.23 8.35 48.26
#>  $ metadata:List of 5
#>   ..$ attribution: chr "openrouteservice.org | OpenStreetMap contributors"
#>   ..$ service    : chr "routing"
#>   ..$ timestamp  : num 1.56e+12
#>   ..$ query      :List of 4
#>   .. ..$ coordinates: num [1:2, 1:2] 8.34 8.34 48.23 48.26
#>   .. ..$ profile    : chr "driving-car"
#>   .. ..$ preference : chr "fastest"
#>   .. ..$ format     : chr "geojson"
#>   ..$ engine     :List of 2
#>   .. ..$ version   : chr "5.0.1"
#>   .. ..$ build_date: chr "2019-05-23T11:54:00Z"

# customized options
ors_directions(coordinates, profile="cycling-mountain", elevation=TRUE)

#> <ors_directions>
#> List of 4
#>  $ type    : chr "FeatureCollection"
#>  $ features:List of 1
#>   ..$ :List of 4
#>   .. ..$ bbox      : num [1:4] 8.33 48.23 8.35 48.26
#>   .. ..$ type      : chr "Feature"
#>   .. ..$ properties:List of 5
#>   .. .. ..$ ascent    : num 166
#>   .. .. ..$ descent   : num 210
#>   .. .. ..$ segments  :List of 1
#>   .. .. .. ..$ :List of 5
#>   .. .. .. .. ..$ distance: num 5510
#>   .. .. .. .. ..$ duration: num 1193
#>   .. .. .. .. ..$ steps   :List of 18
#>   .. .. .. .. .. ..$ :List of 6
#>   .. .. .. .. .. .. ..$ distance   : num 307
#>   .. .. .. .. .. .. ..$ duration   : num 61.5
#>   .. .. .. .. .. .. ..$ type       : int 11
#>   .. .. .. .. .. .. ..$ instruction: chr "Head south on Benatweg"
#>   .. .. .. .. .. .. ..$ name       : chr "Benatweg"
#>   .. .. .. .. .. .. ..$ way_points : int [1:2] 0 12
#>   .. .. .. .. .. ..$ :List of 6
#>   .. .. .. .. .. .. ..$ distance   : num 823
#>   .. .. .. .. .. .. ..$ duration   : num 165
#>   .. .. .. .. .. .. ..$ type       : int 1
#>   .. .. .. .. .. .. ..$ instruction: chr "Turn right onto Hölzle, K 5528"
#>   .. .. .. .. .. .. ..$ name       : chr "Hölzle, K 5528"
#>   .. .. .. .. .. .. ..$ way_points : int [1:2] 12 36
#>   .. .. .. .. .. ..$ :List of 6
#>   .. .. .. .. .. .. ..$ distance   : num 51.7
#>   .. .. .. .. .. .. ..$ duration   : num 10.3
#>   .. .. .. .. .. .. ..$ type       : int 13
#>   .. .. .. .. .. .. ..$ instruction: chr "Keep right onto Sulzbacher Straße, K 5528"
#>   .. .. .. .. .. .. ..$ name       : chr "Sulzbacher Straße, K 5528"
#>   .. .. .. .. .. .. ..$ way_points : int [1:2] 36 39
#>   .. .. .. .. .. ..$ :List of 6
#>   .. .. .. .. .. .. ..$ distance   : num 321
#>   .. .. .. .. .. .. ..$ duration   : num 64.2
#>   .. .. .. .. .. .. ..$ type       : int 6
#>   .. .. .. .. .. .. ..$ instruction: chr "Continue straight onto Sulzbacher Straße, K 5528"
#>   .. .. .. .. .. .. ..$ name       : chr "Sulzbacher Straße, K 5528"
#>   .. .. .. .. .. .. ..$ way_points : int [1:2] 39 48
#>   .. .. .. .. .. ..$ :List of 6
#>   .. .. .. .. .. .. ..$ distance   : num 493
#>   .. .. .. .. .. .. ..$ duration   : num 98.5
#>   .. .. .. .. .. .. ..$ type       : int 4
#>   .. .. .. .. .. .. ..$ instruction: chr "Turn slight left onto Sulzbacher Straße, K 5528"
#>   .. .. .. .. .. .. ..$ name       : chr "Sulzbacher Straße, K 5528"
#>   .. .. .. .. .. .. ..$ way_points : int [1:2] 48 63
#>   .. .. .. .. .. ..$ :List of 6
#>   .. .. .. .. .. .. ..$ distance   : num 265
#>   .. .. .. .. .. .. ..$ duration   : num 59.7
#>   .. .. .. .. .. .. ..$ type       : int 13
#>   .. .. .. .. .. .. ..$ instruction: chr "Keep right"
#>   .. .. .. .. .. .. ..$ name       : chr "-"
#>   .. .. .. .. .. .. ..$ way_points : int [1:2] 63 71
#>   .. .. .. .. .. .. [list output truncated]
#>   .. .. .. .. ..$ ascent  : num 166
#>   .. .. .. .. ..$ descent : num 210
#>   .. .. ..$ summary   :List of 2
#>   .. .. .. ..$ distance: num 5510
#>   .. .. .. ..$ duration: num 1193
#>   .. .. ..$ way_points: int [1:2] 0 164
#>   .. ..$ geometry  :List of 2
#>   .. .. ..$ coordinates: num [1:165, 1:3] 8.34 8.34 8.34 8.34 8.34 ...
#>   .. .. ..$ type       : chr "LineString"
#>  $ bbox    : num [1:4] 8.33 48.23 8.35 48.26
#>  $ metadata:List of 5
#>   ..$ attribution: chr "openrouteservice.org | OpenStreetMap contributors"
#>   ..$ service    : chr "routing"
#>   ..$ timestamp  : num 1.56e+12
#>   ..$ query      :List of 4
#>   .. ..$ coordinates: num [1:2, 1:2] 8.34 8.34 48.23 48.26
#>   .. ..$ profile    : chr "cycling-mountain"
#>   .. ..$ format     : chr "geojson"
#>   .. ..$ elevation  : logi TRUE
#>   ..$ engine     :List of 2
#>   .. ..$ version   : chr "5.0.1"
#>   .. ..$ build_date: chr "2019-05-23T12:01:45Z"

# list of locations as `data.frame` output as simple features `sf` object
locations <- data.frame(lng = c(8.34234, 8.327807, 8.34423),
                        lat = c(48.23424, 48.239368, 48.26424))
ors_directions(locations, output = "sf")

#> Simple feature collection with 1 feature and 3 fields
#> geometry type:  LINESTRING
#> dimension:      XY
#> bbox:           xmin: 8.326189 ymin: 48.23195 xmax: 8.345244 ymax: 48.26355
#> epsg (SRID):    4326
#> proj4string:    +proj=longlat +datum=WGS84 +no_defs
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                              segments
#> 1 2044.5, 329.8, 255, 61.2, 11, Head south on Benatweg, Benatweg, 0, 9, 52.4, 12.6, 6, Continue straight onto Benatweg, Benatweg, 9, 12, 823.2, 65.9, 1, Turn right onto Hölzle, K 5528, Hölzle, K 5528, 12, 36, 790.9, 169.5, 12, Keep left, -, 36, 72, 49.1, 11.8, 4, Turn slight left, -, 72, 73, 73.9, 8.9, 1, Turn right, -, 73, 76, 0, 0, 10, Arrive at your destination, on the right, -, 76, 76, 4287.7, 705.9, 178.6, 21.4, 11, Head northwest, -, 77, 83, 449.8, 54, 13, Keep right, -, 83, 89, 312.9, 46.5, 0, Turn left, -, 89, 100, 6.8, 1.6, 5, Turn slight right onto Sulzbacher Straße, Sulzbacher Straße, 100, 101, 48.8, 7.6, 1, Turn right onto Sulzbacher Straße, K 5528, Sulzbacher Straße, K 5528, 101, 105, 444.5, 53.3, 0, Turn left onto Sulzbacher Straße, Sulzbacher Straße, 105, 118, 436.6, 52.4, 0, Turn left, -, 118, 128, 247.6, 29.7, 13, Keep right, -, 128, 137, 293.7, 52.9, 1, Turn right, -, 137, 147, 252.4, 45.4, 12, Keep left, -, 147, 153, 370.6, 67.6, 13, Keep right, -, 153, 162, 572.1, 111.9, 12, Keep left, -, 162, 178, 37.6, 9, 0, Turn left, -, 178, 181, 469.9, 112.8, 12, Keep left, -, 181, 195, 165.9, 39.8, 2, Turn sharp left, -, 195, 200, 0, 0, 10, Arrive at your destination, on the right, -, 200, 200
#>          summary way_points                       geometry
#> 1 6332.2, 1035.7 0, 76, 199 LINESTRING (8.344268 48.233...