Openrouteservice Geocoding

Resolve input coordinates to addresses and vice versa.

ors_geocode(query, location, ...)

Arguments

query
Name of location, street address or postal code. For a structured geocoding request a named list of parameters.

location
Coordinates to be inquired provided in the form c(longitude, latitude)

...
Optional parameters as described here.

Value

Geocoding: a JSON formatted list of objects corresponding to the search input. Reverse geocoding: the next enclosing object with an address tag which surrounds the given coordinate.

Details

This endpoint can be used for geocoding (specified query) and reverse geocoding requests (specified location). Either query or location has to be specified for a valid request. If both parameters are specified location takes precedence.

Examples

## locations of Heidelberg around the globe
x = ors_geocode("Heidelberg")

## set the number of results returne

x = ors_geocode("Heidelberg", size = 1)

## search within a particular country
ors_geocode("Heidelberg", boundary.country = "DE")

#> <ors_geocode/search>
#> List of 4
#>  $ geocoding:List of 5
#>   ..$ version    : chr "0.2"
#>   ..$ attribution: chr "/v1/attribution"
#>   ..$ query      :List of 6
#>   .. ..$ text            : chr "Heidelberg"
#>   .. ..$ size            : int 10
#>   .. ..$ private         : logi FALSE
#>   .. ..$ boundary.country: chr "DEU"
#>   .. ..$ lang            :List of 4
#>   .. .. ..$ name     : chr "English"
#>   .. .. ..$ iso6391  : chr "en"
#>   .. .. ..$ iso6393  : chr "eng"
#>   .. .. ..$ defaulted: logi TRUE
#>   .. ..$ querySize       : int 20
#>   ..$ engine     :List of 3
#>   .. ..$ name   : chr "Pelias"
#>   .. ..$ author : chr "Mapzen"
#>   .. ..$ version: chr "1.0"
#>   ..$ timestamp  : num 1.52e+12
#>  $ type     : chr "FeatureCollection"
#>  $ features :List of 10
#>   ..$ :List of 4
#>   .. ..$ type      : chr "Feature"
#>   .. ..$ geometry  :List of 2
#>   .. .. ..$ type       : chr "Point"
#>   .. .. ..$ coordinates: num [1:2] 8.68 49.41
#>   .. ..$ properties:List of 20
#>   .. .. ..$ id             : chr "101748539"
#>   .. .. ..$ gid            : chr "whosonfirst:locality:101748539"
#>   .. .. ..$ layer          : chr "locality"
#>   .. .. ..$ source         : chr "whosonfirst"
#>   .. .. ..$ source_id      : chr "101748539"
#>   .. .. ..$ name           : chr "Heidelberg"
#>   .. .. .. [list output truncated]
#>   .. ..$ bbox      : num [1:4] 8.63 49.36 8.78 49.45
#>   ..$ :List of 4
#>   .. ..$ type      : chr "Feature"
#>   .. ..$ geometry  :List of 2
#>   .. .. ..$ type       : chr "Point"
#>   .. .. ..$ coordinates: num [1:2] 8.77 49.38
#>   .. ..$ properties:List of 20
#>   .. .. ..$ id             : chr "101906011"
#>   .. .. ..$ gid            : chr "whosonfirst:locality:101906011"
#>   .. .. ..$ layer          : chr "locality"
#>   .. .. ..$ source         : chr "whosonfirst"
#>   .. .. ..$ source_id      : chr "101906011"
#>   .. .. ..$ name           : chr "Heidelberg"
#>   .. .. .. [list output truncated]
#>   .. ..$ bbox      : num [1:4] 8.76 49.37 8.78 49.38
#>   ..$ :List of 3
#>   .. ..$ type      : chr "Feature"
#>   .. ..$ geometry  :List of 2
#>   .. .. ..$ type       : chr "Point"
#>   .. .. ..$ coordinates: num [1:2] 9.91 52.22
#>   .. ..$ properties:List of 16
#>   .. .. ..$ id         : chr "node:3750741294"
#>   .. .. ..$ gid        : chr "openstreetmap:venue:node:3750741294"
#>   .. .. ..$ layer      : chr "venue"
#>   .. .. ..$ source     : chr "openstreetmap"
#>   .. .. ..$ source_id  : chr "node:3750741294"
#>   .. .. ..$ name       : chr "Heidelberg"
#>   .. .. .. [list output truncated]
#>   ..$ :List of 3
#>   .. ..$ type      : chr "Feature"
#>   .. ..$ geometry  :List of 2
#>   .. .. ..$ type       : chr "Point"
#>   .. .. ..$ coordinates: num [1:2] 9.58 50.65
#>   .. ..$ properties:List of 18
#>   .. .. ..$ id             : chr "node:418877161"
#>   .. .. ..$ gid            : chr "openstreetmap:venue:node:418877161"
#>   .. .. ..$ layer          : chr "venue"
#>   .. .. ..$ source         : chr "openstreetmap"
#>   .. .. ..$ source_id      : chr "node:418877161"
#>   .. .. ..$ name           : chr "Heidelberg"
#>   .. .. .. [list output truncated]
#>   ..$ :List of 3
#>   .. ..$ type      : chr "Feature"
#>   .. ..$ geometry  :List of 2
#>   .. .. ..$ type       : chr "Point"
#>   .. .. ..$ coordinates: num [1:2] 8.8 49.7
#>   .. ..$ properties:List of 18
#>   .. .. ..$ id             : chr "node:393965464"
#>   .. .. ..$ gid            : chr "openstreetmap:venue:node:393965464"
#>   .. .. ..$ layer          : chr "venue"
#>   .. .. ..$ source         : chr "openstreetmap"
#>   .. .. ..$ source_id      : chr "node:393965464"
#>   .. .. ..$ name           : chr "Heidelberg"
#>   .. .. .. [list output truncated]
#>   ..$ :List of 3
#>   .. ..$ type      : chr "Feature"
#>   .. ..$ geometry  :List of 2
#>   .. .. ..$ type       : chr "Point"
#>   .. .. ..$ coordinates: num [1:2] 6.94 49.72
#>   .. ..$ properties:List of 16
#>   .. .. ..$ id         : chr "node:4317353171"
#>   .. .. ..$ gid        : chr "openstreetmap:venue:node:4317353171"
#>   .. .. ..$ layer      : chr "venue"
#>   .. .. ..$ source     : chr "openstreetmap"
#>   .. .. ..$ source_id  : chr "node:4317353171"
#>   .. .. ..$ name       : chr "Heidelberg"
#>   .. .. .. [list output truncated]
#>   .. [list output truncated]
#>  $ bbox     : num [1:4] 6.94 49.11 13.11 52.22

## structured geocoding
x = ors_geocode(list(locality="Heidelberg", county="Heidelberg"))

## reverse geocoding
location = x$features[[1L]]$geometry$coordinates
y = ors_geocode(location = location, layers = "locality", size = 1)