Openrouteservice Matrix

Obtain one-to-many, many-to-one and many-to-many matrices for time and distance.

ors_matrix(
  locations,
  profile = ors_profile(),
  ...,
  api_key = ors_api_key(),
  output = c("parsed", "text")
)

Arguments

locations
List of longitude, latitude coordinate pairs, alternatively a two column matrix or data.frame.

profile
Route profile, defaults to "driving-car".

...
Optional parameters as described here

api_key
Character scalar containing openrouteservice API key

output
Output format. By default the response is being parsed to a list-based R object

Value

Duration or distance matrix for multiple source and destination

for "text", a character vector of length 1 re-encoded to UTF-8.

for "parsed", a parsed R object.

Examples

coordinates = list(
  c(9.970093, 48.477473),
  c(9.207916, 49.153868),
  c(37.573242, 55.801281),
  c(115.663757,38.106467)
)

# query for duration and distance in km
res = ors_matrix(coordinates, metrics = c("duration", "distance"), units = "km")

# duration in hours
res$durations / 3600

#>            [,1]       [,2]     [,3]      [,4]
#> [1,]   0.000000   1.655122 25.04572 111.16592
#> [2,]   1.674261   0.000000 24.75677 110.87698
#> [3,]  24.841267  24.522764  0.00000  87.00542
#> [4,] 111.094689 110.776189 86.78183   0.00000

# distance in km
res$distances

#>          [,1]     [,2]    [,3]     [,4]
#> [1,]     0.00   155.07 2354.46 10187.32
#> [2,]   155.68     0.00 2336.34 10169.20
#> [3,]  2410.39  2383.03    0.00  7628.78
#> [4,] 10250.85 10223.49 7635.39     0.00