Openrouteservice POIs
Source: R/pois.R

Search for points of interest around points or in geometries.

ors_pois(request = c("pois", "stats", "list"), geometry, ...)

Arguments

request
One of the following: "pois", "stats" or "list"

geometry
GeoJSON geometry object (Point, Linestring or Polygon)

...
Optional request attributes as described here.

Value

A list of points of interest in the area specified in geometry.

Details

There are three different request types: pois, stats and list.

pois returns a GeoJSON FeatureCollection in the specified bounding box or geometry. stats does the same but groups by categories, ultimately returning a JSON object with the absolute numbers of POIs of a certain group.

list returns a list of category groups and their ids.

Examples

ors_pois('list')

#> No encoding supplied: defaulting to UTF-8.

#> <ors_pois>
#> List of 16
#>  $ accomodation             :List of 2
#>   ..$ children:List of 1
#>   .. ..$ tourism:List of 20
#>   .. .. ..$ alpine_hut    : int 101
#>   .. .. ..$ appartment    : int 102
#>   .. .. ..$ aquarium      : int 308
#>   .. .. ..$ artwork       : int 621
#>   .. .. ..$ attraction    : int 622
#>   .. .. ..$ camp_site     : int 103
#>   .. .. .. [list output truncated]
#>   ..$ id      : int 100
#>  $ animals                  :List of 2
#>   ..$ children:List of 2
#>   .. ..$ amenity:List of 103
#>   .. .. ..$ animal_boarding       : int 121
#>   .. .. ..$ animal_shelter        : int 122
#>   .. .. ..$ arts_centre           : int 131
#>   .. .. ..$ atm                   : int 191
#>   .. .. ..$ baby_hatch            : int 201
#>   .. .. ..$ bank                  : int 192
#>   .. .. .. [list output truncated]
#>   .. ..$ shop   :List of 123
#>   .. .. ..$ agrarian             : int 421
#>   .. .. ..$ alkohol              : int 422
#>   .. .. ..$ antiques             : int 423
#>   .. .. ..$ art                  : int 424
#>   .. .. ..$ bag                  : int 425
#>   .. .. ..$ bakery               : int 426
#>   .. .. .. [list output truncated]
#>   ..$ id      : int 120
#>  $ arts_and_culture         :List of 2
#>   ..$ children:List of 2
#>   .. ..$ amenity:List of 4
#>   .. .. ..$ arts_centre     : int 131
#>   .. .. ..$ library         : int 133
#>   .. .. ..$ place_of_worship: int 135
#>   .. .. ..$ studio          : int 136
#>   .. ..$ tourism:List of 2
#>   .. .. ..$ gallery: int 132
#>   .. .. ..$ museum : int 134
#>   ..$ id      : int 130
#>  $ education                :List of 2
#>   ..$ children:List of 1
#>   .. ..$ amenity:List of 7
#>   .. .. ..$ college       : int 151
#>   .. .. ..$ driving_school: int 152
#>   .. .. ..$ kindergarten  : int 153
#>   .. .. ..$ lanuage_school: int 154
#>   .. .. ..$ music_school  : int 155
#>   .. .. ..$ school        : int 156
#>   .. .. .. [list output truncated]
#>   ..$ id      : int 150
#>  $ facilities               :List of 2
#>   ..$ children:List of 2
#>   .. ..$ amenity  :List of 20
#>   .. .. ..$ bench                : int 162
#>   .. .. ..$ clock                : int 164
#>   .. .. ..$ compressed_air       : int 161
#>   .. .. ..$ drinking_water       : int 166
#>   .. .. ..$ emergency_phone      : int 163
#>   .. .. ..$ hunting_stand        : int 168
#>   .. .. .. [list output truncated]
#>   .. ..$ emergency:List of 2
#>   .. .. ..$ defibrillator: int 165
#>   .. .. ..$ fire_hydrant : int 167
#>   ..$ id      : int 160
#>  $ financial                :List of 2
#>   ..$ children:List of 1
#>   .. ..$ amenity:List of 3
#>   .. .. ..$ atm             : int 191
#>   .. .. ..$ bank            : int 192
#>   .. .. ..$ bureau_de_change: int 193
#>   ..$ id      : int 190
#>   [list output truncated]