Openrouteservice POIs

Search for points of interest around points or in geometries.

ors_pois(
  request = c("pois", "stats", "list"),
  geometry,
  ...,
  api_key = ors_api_key(),
  output = c("parsed", "text", "sf")
)

Arguments

request

One of the following: "pois", "stats" or "list"

geometry

named list containing either a geojson geometry object (GeoJSON Point, LineString or Polygon) or a bbox, optionally buffered by a value provided buffer

...

Optional request attributes as described here

api_key

Character scalar containing openrouteservice API key

output

Output format. By default the response is being parsed to a list-based R object

Value

A list of points of interest in the area specified in geometry structured according to output:

for "text", a character vector of length 1 re-encoded to UTF-8.

for "parsed", a parsed R object.

for "sf", a simple features sf object. Valid only for argument request = "pois".

Details

There are three different request types: pois, stats and list.

pois returns a GeoJSON FeatureCollection in the bounding box specified in geometry$bbox or a GeoJSON geometry provided in geometry$geojson. stats does the same but groups by categories, ultimately returning a JSON object with the absolute numbers of POIs of a certain group.

list returns a list of category groups and their ids.

Author

Andrzej Oleś andrzej.oles@gmail.com

Examples

# POI categories list
ors_pois('list')
#> <ors_pois>
#> List of 16
#>  $ accomodation             :List of 2
#>   ..$ id      : int 100
#>   ..$ children:List of 1
#>   .. ..$ tourism:List of 10
#>   .. .. ..$ alpine_hut    : int 101
#>   .. .. ..$ apartment     : int 102
#>   .. .. ..$ camp_site     : int 103
#>   .. .. ..$ caravan_site  : int 104
#>   .. .. ..$ chalet        : int 105
#>   .. .. ..$ guest_house   : int 106
#>   .. .. .. [list output truncated]
#>  $ animals                  :List of 2
#>   ..$ id      : int 120
#>   ..$ children:List of 2
#>   .. ..$ amenity:List of 3
#>   .. .. ..$ animal_boarding: int 121
#>   .. .. ..$ animal_shelter : int 122
#>   .. .. ..$ veterinary     : int 123
#>   .. ..$ shop   :List of 1
#>   .. .. ..$ pet: int 124
#>  $ arts_and_culture         :List of 2
#>   ..$ id      : int 130
#>   ..$ children:List of 2
#>   .. ..$ amenity:List of 4
#>   .. .. ..$ arts_centre     : int 131
#>   .. .. ..$ library         : int 133
#>   .. .. ..$ place_of_worship: int 135
#>   .. .. ..$ studio          : int 136
#>   .. ..$ tourism:List of 2
#>   .. .. ..$ gallery: int 132
#>   .. .. ..$ museum : int 134
#>  $ education                :List of 2
#>   ..$ id      : int 150
#>   ..$ children:List of 1
#>   .. ..$ amenity:List of 7
#>   .. .. ..$ college        : int 151
#>   .. .. ..$ driving_school : int 152
#>   .. .. ..$ kindergarten   : int 153
#>   .. .. ..$ language_school: int 154
#>   .. .. ..$ music_school   : int 155
#>   .. .. ..$ school         : int 156
#>   .. .. .. [list output truncated]
#>  $ facilities               :List of 2
#>   ..$ id      : int 160
#>   ..$ children:List of 2
#>   .. ..$ amenity  :List of 19
#>   .. .. ..$ compressed_air       : int 161
#>   .. .. ..$ bench                : int 162
#>   .. .. ..$ emergency_phone      : int 163
#>   .. .. ..$ clock                : int 164
#>   .. .. ..$ drinking_water       : int 166
#>   .. .. ..$ hunting_stand        : int 168
#>   .. .. .. [list output truncated]
#>   .. ..$ emergency:List of 3
#>   .. .. ..$ access_point : int 205
#>   .. .. ..$ defibrillator: int 165
#>   .. .. ..$ fire_hydrant : int 167
#>  $ financial                :List of 2
#>   ..$ id      : int 190
#>   ..$ children:List of 1
#>   .. ..$ amenity:List of 3
#>   .. .. ..$ atm             : int 191
#>   .. .. ..$ bank            : int 192
#>   .. .. ..$ bureau_de_change: int 193
#>   [list output truncated]

# POIs around a buffered point
geometry <- list(geojson = list(type = "Point",
                                coordinates = c(8.8034, 53.0756)),
                buffer = 100)
ors_pois(geometry = geometry)
#> <ors_pois>
#> List of 4
#>  $ type       : chr "FeatureCollection"
#>  $ bbox       : num [1:4] 8.8 53.1 8.8 53.1
#>  $ features   :List of 48
#>   ..$ :List of 3
#>   .. ..$ type      : chr "Feature"
#>   .. ..$ geometry  :List of 2
#>   .. .. ..$ type       : chr "Point"
#>   .. .. ..$ coordinates: num [1:2] 8.8 53.1
#>   .. ..$ properties:List of 4
#>   .. .. ..$ osm_id      : int 27403586
#>   .. .. ..$ osm_type    : int 2
#>   .. .. ..$ distance    : num 22.5
#>   .. .. ..$ category_ids:List of 1
#>   .. .. .. ..$ 601:List of 2
#>   .. .. .. .. ..$ category_name : chr "parking"
#>   .. .. .. .. ..$ category_group: chr "transport"
#>   ..$ :List of 3
#>   .. ..$ type      : chr "Feature"
#>   .. ..$ geometry  :List of 2
#>   .. .. ..$ type       : chr "Point"
#>   .. .. ..$ coordinates: num [1:2] 8.8 53.1
#>   .. ..$ properties:List of 5
#>   .. .. ..$ osm_id      : int 158546065
#>   .. .. ..$ osm_type    : int 2
#>   .. .. ..$ distance    : num 57.1
#>   .. .. ..$ category_ids:List of 2
#>   .. .. .. ..$ 245:List of 2
#>   .. .. .. .. ..$ category_name : chr "ship"
#>   .. .. .. .. ..$ category_group: chr "historic"
#>   .. .. .. ..$ 622:List of 2
#>   .. .. .. .. ..$ category_name : chr "attraction"
#>   .. .. .. .. ..$ category_group: chr "tourism"
#>   .. .. ..$ osm_tags    :List of 3
#>   .. .. .. ..$ opening_hours: chr "Su-Fr 12:00-20:00; Sa 11:00-21:00"
#>   .. .. .. ..$ website      : chr "https://admiral-nelson.de"
#>   .. .. .. ..$ name         : chr "Admiral Nelson"
#>   ..$ :List of 3
#>   .. ..$ type      : chr "Feature"
#>   .. ..$ geometry  :List of 2
#>   .. .. ..$ type       : chr "Point"
#>   .. .. ..$ coordinates: num [1:2] 8.8 53.1
#>   .. ..$ properties:List of 5
#>   .. .. ..$ osm_id      : int 163626811
#>   .. .. ..$ osm_type    : int 2
#>   .. .. ..$ distance    : num 88.7
#>   .. .. ..$ category_ids:List of 1
#>   .. .. .. ..$ 135:List of 2
#>   .. .. .. .. ..$ category_name : chr "place_of_worship"
#>   .. .. .. .. ..$ category_group: chr "arts_and_culture"
#>   .. .. ..$ osm_tags    :List of 3
#>   .. .. .. ..$ name         : chr "Martinikirche"
#>   .. .. .. ..$ website      : chr "https://www.st-martini.net/"
#>   .. .. .. ..$ opening_hours: chr "May-Oct: 10:30-12:30"
#>   ..$ :List of 3
#>   .. ..$ type      : chr "Feature"
#>   .. ..$ geometry  :List of 2
#>   .. .. ..$ type       : chr "Point"
#>   .. .. ..$ coordinates: num [1:2] 8.8 53.1
#>   .. ..$ properties:List of 5
#>   .. .. ..$ osm_id      : int 235754690
#>   .. .. ..$ osm_type    : int 2
#>   .. .. ..$ distance    : num 91.7
#>   .. .. ..$ category_ids:List of 1
#>   .. .. .. ..$ 237:List of 2
#>   .. .. .. .. ..$ category_name : chr "memorial"
#>   .. .. .. .. ..$ category_group: chr "historic"
#>   .. .. ..$ osm_tags    :List of 1
#>   .. .. .. ..$ name: chr "Balge"
#>   ..$ :List of 3
#>   .. ..$ type      : chr "Feature"
#>   .. ..$ geometry  :List of 2
#>   .. .. ..$ type       : chr "Point"
#>   .. .. ..$ coordinates: num [1:2] 8.8 53.1
#>   .. ..$ properties:List of 5
#>   .. .. ..$ osm_id      : int 339719410
#>   .. .. ..$ osm_type    : int 2
#>   .. .. ..$ distance    : num 85
#>   .. .. ..$ category_ids:List of 2
#>   .. .. .. ..$ 245:List of 2
#>   .. .. .. .. ..$ category_name : chr "ship"
#>   .. .. .. .. ..$ category_group: chr "historic"
#>   .. .. .. ..$ 622:List of 2
#>   .. .. .. .. ..$ category_name : chr "attraction"
#>   .. .. .. .. ..$ category_group: chr "tourism"
#>   .. .. ..$ osm_tags    :List of 3
#>   .. .. .. ..$ name         : chr "Alexander von Humboldt"
#>   .. .. .. ..$ opening_hours: chr "Mo-Su,PH 11:30+"
#>   .. .. .. ..$ website      : chr "https://alex-das-schiff.de/"
#>   ..$ :List of 3
#>   .. ..$ type      : chr "Feature"
#>   .. ..$ geometry  :List of 2
#>   .. .. ..$ type       : chr "Point"
#>   .. .. ..$ coordinates: num [1:2] 8.8 53.1
#>   .. ..$ properties:List of 5
#>   .. .. ..$ osm_id      : int 458181680
#>   .. .. ..$ osm_type    : int 1
#>   .. .. ..$ distance    : num 24.8
#>   .. .. ..$ category_ids:List of 1
#>   .. .. .. ..$ 570:List of 2
#>   .. .. .. .. ..$ category_name : chr "restaurant"
#>   .. .. .. .. ..$ category_group: chr "sustenance"
#>   .. .. ..$ osm_tags    :List of 5
#>   .. .. .. ..$ name         : chr "Osteria"
#>   .. .. .. ..$ wheelchair   : chr "no"
#>   .. .. .. ..$ website      : chr "http://www.osteria-bremen.de/"
#>   .. .. .. ..$ phone        : chr "+49 421 3398207"
#>   .. .. .. ..$ opening_hours: chr "Mo-Sa 12:00-15:00,18:00-23:00, Su 18:00+"
#>   .. [list output truncated]
#>  $ information:List of 4
#>   ..$ attribution: chr "openrouteservice.org | OpenStreetMap contributors"
#>   ..$ version    : chr "0.1"
#>   ..$ timestamp  : int 1727437217
#>   ..$ query      :List of 2
#>   .. ..$ request : chr "pois"
#>   .. ..$ geometry:List of 2
#>   .. .. ..$ geojson:List of 2
#>   .. .. .. ..$ type       : chr "Point"
#>   .. .. .. ..$ coordinates: num [1:2] 8.8 53.1
#>   .. .. ..$ buffer : int 100

# alternative specification via bounding box
ors_pois(geometry = list(bbox = list(c(8.8034, 53.0756), c(8.8034, 53.0756)),
                         buffer = 100))
#> <ors_pois>
#> List of 4
#>  $ type       : chr "FeatureCollection"
#>  $ bbox       : num [1:4] 8.8 53.1 8.8 53.1
#>  $ features   :List of 48
#>   ..$ :List of 3
#>   .. ..$ type      : chr "Feature"
#>   .. ..$ geometry  :List of 2
#>   .. .. ..$ type       : chr "Point"
#>   .. .. ..$ coordinates: num [1:2] 8.8 53.1
#>   .. ..$ properties:List of 4
#>   .. .. ..$ osm_id      : int 27403586
#>   .. .. ..$ osm_type    : int 2
#>   .. .. ..$ distance    : num 22.5
#>   .. .. ..$ category_ids:List of 1
#>   .. .. .. ..$ 601:List of 2
#>   .. .. .. .. ..$ category_name : chr "parking"
#>   .. .. .. .. ..$ category_group: chr "transport"
#>   ..$ :List of 3
#>   .. ..$ type      : chr "Feature"
#>   .. ..$ geometry  :List of 2
#>   .. .. ..$ type       : chr "Point"
#>   .. .. ..$ coordinates: num [1:2] 8.8 53.1
#>   .. ..$ properties:List of 5
#>   .. .. ..$ osm_id      : int 158546065
#>   .. .. ..$ osm_type    : int 2
#>   .. .. ..$ distance    : num 57.1
#>   .. .. ..$ category_ids:List of 2
#>   .. .. .. ..$ 245:List of 2
#>   .. .. .. .. ..$ category_name : chr "ship"
#>   .. .. .. .. ..$ category_group: chr "historic"
#>   .. .. .. ..$ 622:List of 2
#>   .. .. .. .. ..$ category_name : chr "attraction"
#>   .. .. .. .. ..$ category_group: chr "tourism"
#>   .. .. ..$ osm_tags    :List of 3
#>   .. .. .. ..$ opening_hours: chr "Su-Fr 12:00-20:00; Sa 11:00-21:00"
#>   .. .. .. ..$ website      : chr "https://admiral-nelson.de"
#>   .. .. .. ..$ name         : chr "Admiral Nelson"
#>   ..$ :List of 3
#>   .. ..$ type      : chr "Feature"
#>   .. ..$ geometry  :List of 2
#>   .. .. ..$ type       : chr "Point"
#>   .. .. ..$ coordinates: num [1:2] 8.8 53.1
#>   .. ..$ properties:List of 5
#>   .. .. ..$ osm_id      : int 163626811
#>   .. .. ..$ osm_type    : int 2
#>   .. .. ..$ distance    : num 88.7
#>   .. .. ..$ category_ids:List of 1
#>   .. .. .. ..$ 135:List of 2
#>   .. .. .. .. ..$ category_name : chr "place_of_worship"
#>   .. .. .. .. ..$ category_group: chr "arts_and_culture"
#>   .. .. ..$ osm_tags    :List of 3
#>   .. .. .. ..$ name         : chr "Martinikirche"
#>   .. .. .. ..$ website      : chr "https://www.st-martini.net/"
#>   .. .. .. ..$ opening_hours: chr "May-Oct: 10:30-12:30"
#>   ..$ :List of 3
#>   .. ..$ type      : chr "Feature"
#>   .. ..$ geometry  :List of 2
#>   .. .. ..$ type       : chr "Point"
#>   .. .. ..$ coordinates: num [1:2] 8.8 53.1
#>   .. ..$ properties:List of 5
#>   .. .. ..$ osm_id      : int 235754690
#>   .. .. ..$ osm_type    : int 2
#>   .. .. ..$ distance    : num 91.7
#>   .. .. ..$ category_ids:List of 1
#>   .. .. .. ..$ 237:List of 2
#>   .. .. .. .. ..$ category_name : chr "memorial"
#>   .. .. .. .. ..$ category_group: chr "historic"
#>   .. .. ..$ osm_tags    :List of 1
#>   .. .. .. ..$ name: chr "Balge"
#>   ..$ :List of 3
#>   .. ..$ type      : chr "Feature"
#>   .. ..$ geometry  :List of 2
#>   .. .. ..$ type       : chr "Point"
#>   .. .. ..$ coordinates: num [1:2] 8.8 53.1
#>   .. ..$ properties:List of 5
#>   .. .. ..$ osm_id      : int 339719410
#>   .. .. ..$ osm_type    : int 2
#>   .. .. ..$ distance    : num 85
#>   .. .. ..$ category_ids:List of 2
#>   .. .. .. ..$ 245:List of 2
#>   .. .. .. .. ..$ category_name : chr "ship"
#>   .. .. .. .. ..$ category_group: chr "historic"
#>   .. .. .. ..$ 622:List of 2
#>   .. .. .. .. ..$ category_name : chr "attraction"
#>   .. .. .. .. ..$ category_group: chr "tourism"
#>   .. .. ..$ osm_tags    :List of 3
#>   .. .. .. ..$ name         : chr "Alexander von Humboldt"
#>   .. .. .. ..$ opening_hours: chr "Mo-Su,PH 11:30+"
#>   .. .. .. ..$ website      : chr "https://alex-das-schiff.de/"
#>   ..$ :List of 3
#>   .. ..$ type      : chr "Feature"
#>   .. ..$ geometry  :List of 2
#>   .. .. ..$ type       : chr "Point"
#>   .. .. ..$ coordinates: num [1:2] 8.8 53.1
#>   .. ..$ properties:List of 5
#>   .. .. ..$ osm_id      : int 458181680
#>   .. .. ..$ osm_type    : int 1
#>   .. .. ..$ distance    : num 24.8
#>   .. .. ..$ category_ids:List of 1
#>   .. .. .. ..$ 570:List of 2
#>   .. .. .. .. ..$ category_name : chr "restaurant"
#>   .. .. .. .. ..$ category_group: chr "sustenance"
#>   .. .. ..$ osm_tags    :List of 5
#>   .. .. .. ..$ name         : chr "Osteria"
#>   .. .. .. ..$ wheelchair   : chr "no"
#>   .. .. .. ..$ website      : chr "http://www.osteria-bremen.de/"
#>   .. .. .. ..$ phone        : chr "+49 421 3398207"
#>   .. .. .. ..$ opening_hours: chr "Mo-Sa 12:00-15:00,18:00-23:00, Su 18:00+"
#>   .. [list output truncated]
#>  $ information:List of 4
#>   ..$ attribution: chr "openrouteservice.org | OpenStreetMap contributors"
#>   ..$ version    : chr "0.1"
#>   ..$ timestamp  : int 1727437217
#>   ..$ query      :List of 2
#>   .. ..$ request : chr "pois"
#>   .. ..$ geometry:List of 2
#>   .. .. ..$ bbox  : num [1:2, 1:2] 8.8 8.8 53.1 53.1
#>   .. .. ..$ buffer: int 100

# POIs of given categories
ors_pois(geometry = geometry,
         limit = 200,
         sortby = "distance",
         filters = list(
           category_ids = c(180, 245)
         ))
#> <ors_pois>
#> List of 4
#>  $ type       : chr "FeatureCollection"
#>  $ bbox       : num [1:4] 8.8 53.1 8.8 53.1
#>  $ features   :List of 9
#>   ..$ :List of 3
#>   .. ..$ type      : chr "Feature"
#>   .. ..$ geometry  :List of 2
#>   .. .. ..$ type       : chr "Point"
#>   .. .. ..$ coordinates: num [1:2] 8.8 53.1
#>   .. ..$ properties:List of 4
#>   .. .. ..$ osm_id      : num 4.99e+09
#>   .. .. ..$ osm_type    : int 1
#>   .. .. ..$ distance    : num 20.9
#>   .. .. ..$ category_ids:List of 1
#>   .. .. .. ..$ 180:List of 2
#>   .. .. .. .. ..$ category_name : chr "waste_basket"
#>   .. .. .. .. ..$ category_group: chr "facilities"
#>   ..$ :List of 3
#>   .. ..$ type      : chr "Feature"
#>   .. ..$ geometry  :List of 2
#>   .. .. ..$ type       : chr "Point"
#>   .. .. ..$ coordinates: num [1:2] 8.8 53.1
#>   .. ..$ properties:List of 4
#>   .. .. ..$ osm_id      : num 4.99e+09
#>   .. .. ..$ osm_type    : int 1
#>   .. .. ..$ distance    : num 21.8
#>   .. .. ..$ category_ids:List of 1
#>   .. .. .. ..$ 180:List of 2
#>   .. .. .. .. ..$ category_name : chr "waste_basket"
#>   .. .. .. .. ..$ category_group: chr "facilities"
#>   ..$ :List of 3
#>   .. ..$ type      : chr "Feature"
#>   .. ..$ geometry  :List of 2
#>   .. .. ..$ type       : chr "Point"
#>   .. .. ..$ coordinates: num [1:2] 8.8 53.1
#>   .. ..$ properties:List of 4
#>   .. .. ..$ osm_id      : num 4.99e+09
#>   .. .. ..$ osm_type    : int 1
#>   .. .. ..$ distance    : num 26.2
#>   .. .. ..$ category_ids:List of 1
#>   .. .. .. ..$ 180:List of 2
#>   .. .. .. .. ..$ category_name : chr "waste_basket"
#>   .. .. .. .. ..$ category_group: chr "facilities"
#>   ..$ :List of 3
#>   .. ..$ type      : chr "Feature"
#>   .. ..$ geometry  :List of 2
#>   .. .. ..$ type       : chr "Point"
#>   .. .. ..$ coordinates: num [1:2] 8.8 53.1
#>   .. ..$ properties:List of 4
#>   .. .. ..$ osm_id      : num 4.99e+09
#>   .. .. ..$ osm_type    : int 1
#>   .. .. ..$ distance    : num 34.8
#>   .. .. ..$ category_ids:List of 1
#>   .. .. .. ..$ 180:List of 2
#>   .. .. .. .. ..$ category_name : chr "waste_basket"
#>   .. .. .. .. ..$ category_group: chr "facilities"
#>   ..$ :List of 3
#>   .. ..$ type      : chr "Feature"
#>   .. ..$ geometry  :List of 2
#>   .. .. ..$ type       : chr "Point"
#>   .. .. ..$ coordinates: num [1:2] 8.8 53.1
#>   .. ..$ properties:List of 4
#>   .. .. ..$ osm_id      : num 4.99e+09
#>   .. .. ..$ osm_type    : int 1
#>   .. .. ..$ distance    : num 45.1
#>   .. .. ..$ category_ids:List of 1
#>   .. .. .. ..$ 180:List of 2
#>   .. .. .. .. ..$ category_name : chr "waste_basket"
#>   .. .. .. .. ..$ category_group: chr "facilities"
#>   ..$ :List of 3
#>   .. ..$ type      : chr "Feature"
#>   .. ..$ geometry  :List of 2
#>   .. .. ..$ type       : chr "Point"
#>   .. .. ..$ coordinates: num [1:2] 8.8 53.1
#>   .. ..$ properties:List of 5
#>   .. .. ..$ osm_id      : int 158546065
#>   .. .. ..$ osm_type    : int 2
#>   .. .. ..$ distance    : num 57.1
#>   .. .. ..$ category_ids:List of 1
#>   .. .. .. ..$ 245:List of 2
#>   .. .. .. .. ..$ category_name : chr "ship"
#>   .. .. .. .. ..$ category_group: chr "historic"
#>   .. .. ..$ osm_tags    :List of 3
#>   .. .. .. ..$ name         : chr "Admiral Nelson"
#>   .. .. .. ..$ opening_hours: chr "Su-Fr 12:00-20:00; Sa 11:00-21:00"
#>   .. .. .. ..$ website      : chr "https://admiral-nelson.de"
#>   .. [list output truncated]
#>  $ information:List of 4
#>   ..$ attribution: chr "openrouteservice.org | OpenStreetMap contributors"
#>   ..$ version    : chr "0.1"
#>   ..$ timestamp  : int 1727437218
#>   ..$ query      :List of 5
#>   .. ..$ request : chr "pois"
#>   .. ..$ limit   : int 200
#>   .. ..$ sortby  : chr "distance"
#>   .. ..$ filters :List of 1
#>   .. .. ..$ category_ids: int [1:2] 180 245
#>   .. ..$ geometry:List of 2
#>   .. .. ..$ geojson:List of 2
#>   .. .. .. ..$ type       : chr "Point"
#>   .. .. .. ..$ coordinates: num [1:2] 8.8 53.1
#>   .. .. ..$ buffer : int 100

# POIs of given category groups
ors_pois(geometry = geometry,
         limit = 200,
         sortby = "distance",
         filters = list(
           category_group_ids = 160
         ))
#> <ors_pois>
#> List of 4
#>  $ type       : chr "FeatureCollection"
#>  $ bbox       : num [1:4] 8.8 53.1 8.8 53.1
#>  $ features   :List of 21
#>   ..$ :List of 3
#>   .. ..$ type      : chr "Feature"
#>   .. ..$ geometry  :List of 2
#>   .. .. ..$ type       : chr "Point"
#>   .. .. ..$ coordinates: num [1:2] 8.8 53.1
#>   .. ..$ properties:List of 4
#>   .. .. ..$ osm_id      : num 4.99e+09
#>   .. .. ..$ osm_type    : int 1
#>   .. .. ..$ distance    : num 20.6
#>   .. .. ..$ category_ids:List of 1
#>   .. .. .. ..$ 162:List of 2
#>   .. .. .. .. ..$ category_name : chr "bench"
#>   .. .. .. .. ..$ category_group: chr "facilities"
#>   ..$ :List of 3
#>   .. ..$ type      : chr "Feature"
#>   .. ..$ geometry  :List of 2
#>   .. .. ..$ type       : chr "Point"
#>   .. .. ..$ coordinates: num [1:2] 8.8 53.1
#>   .. ..$ properties:List of 4
#>   .. .. ..$ osm_id      : num 4.99e+09
#>   .. .. ..$ osm_type    : int 1
#>   .. .. ..$ distance    : num 20.9
#>   .. .. ..$ category_ids:List of 1
#>   .. .. .. ..$ 180:List of 2
#>   .. .. .. .. ..$ category_name : chr "waste_basket"
#>   .. .. .. .. ..$ category_group: chr "facilities"
#>   ..$ :List of 3
#>   .. ..$ type      : chr "Feature"
#>   .. ..$ geometry  :List of 2
#>   .. .. ..$ type       : chr "Point"
#>   .. .. ..$ coordinates: num [1:2] 8.8 53.1
#>   .. ..$ properties:List of 4
#>   .. .. ..$ osm_id      : num 4.99e+09
#>   .. .. ..$ osm_type    : int 1
#>   .. .. ..$ distance    : num 21.2
#>   .. .. ..$ category_ids:List of 1
#>   .. .. .. ..$ 162:List of 2
#>   .. .. .. .. ..$ category_name : chr "bench"
#>   .. .. .. .. ..$ category_group: chr "facilities"
#>   ..$ :List of 3
#>   .. ..$ type      : chr "Feature"
#>   .. ..$ geometry  :List of 2
#>   .. .. ..$ type       : chr "Point"
#>   .. .. ..$ coordinates: num [1:2] 8.8 53.1
#>   .. ..$ properties:List of 4
#>   .. .. ..$ osm_id      : num 4.99e+09
#>   .. .. ..$ osm_type    : int 1
#>   .. .. ..$ distance    : num 21.6
#>   .. .. ..$ category_ids:List of 1
#>   .. .. .. ..$ 162:List of 2
#>   .. .. .. .. ..$ category_name : chr "bench"
#>   .. .. .. .. ..$ category_group: chr "facilities"
#>   ..$ :List of 3
#>   .. ..$ type      : chr "Feature"
#>   .. ..$ geometry  :List of 2
#>   .. .. ..$ type       : chr "Point"
#>   .. .. ..$ coordinates: num [1:2] 8.8 53.1
#>   .. ..$ properties:List of 4
#>   .. .. ..$ osm_id      : num 4.99e+09
#>   .. .. ..$ osm_type    : int 1
#>   .. .. ..$ distance    : num 21.8
#>   .. .. ..$ category_ids:List of 1
#>   .. .. .. ..$ 180:List of 2
#>   .. .. .. .. ..$ category_name : chr "waste_basket"
#>   .. .. .. .. ..$ category_group: chr "facilities"
#>   ..$ :List of 3
#>   .. ..$ type      : chr "Feature"
#>   .. ..$ geometry  :List of 2
#>   .. .. ..$ type       : chr "Point"
#>   .. .. ..$ coordinates: num [1:2] 8.8 53.1
#>   .. ..$ properties:List of 4
#>   .. .. ..$ osm_id      : num 4.99e+09
#>   .. .. ..$ osm_type    : int 1
#>   .. .. ..$ distance    : num 22.5
#>   .. .. ..$ category_ids:List of 1
#>   .. .. .. ..$ 162:List of 2
#>   .. .. .. .. ..$ category_name : chr "bench"
#>   .. .. .. .. ..$ category_group: chr "facilities"
#>   .. [list output truncated]
#>  $ information:List of 4
#>   ..$ attribution: chr "openrouteservice.org | OpenStreetMap contributors"
#>   ..$ version    : chr "0.1"
#>   ..$ timestamp  : int 1727437218
#>   ..$ query      :List of 5
#>   .. ..$ request : chr "pois"
#>   .. ..$ limit   : int 200
#>   .. ..$ sortby  : chr "distance"
#>   .. ..$ filters :List of 1
#>   .. .. ..$ category_group_ids: int 160
#>   .. ..$ geometry:List of 2
#>   .. .. ..$ geojson:List of 2
#>   .. .. .. ..$ type       : chr "Point"
#>   .. .. .. ..$ coordinates: num [1:2] 8.8 53.1
#>   .. .. ..$ buffer : int 100

# POI Statistics
ors_pois("stats", geometry = geometry)
#> <ors_pois>
#> List of 2
#>  $ places     :List of 11
#>   ..$ total_count     : int 50
#>   ..$ accomodation    :List of 3
#>   .. ..$ group_id   : int 100
#>   .. ..$ categories :List of 1
#>   .. .. ..$ hotel:List of 2
#>   .. .. .. ..$ count      : int 1
#>   .. .. .. ..$ category_id: int 108
#>   .. ..$ total_count: int 1
#>   ..$ arts_and_culture:List of 3
#>   .. ..$ group_id   : int 130
#>   .. ..$ categories :List of 1
#>   .. .. ..$ place_of_worship:List of 2
#>   .. .. .. ..$ count      : int 1
#>   .. .. .. ..$ category_id: int 135
#>   .. ..$ total_count: int 1
#>   ..$ facilities      :List of 3
#>   .. ..$ group_id   : int 160
#>   .. ..$ categories :List of 3
#>   .. .. ..$ bench       :List of 2
#>   .. .. .. ..$ count      : int 13
#>   .. .. .. ..$ category_id: int 162
#>   .. .. ..$ toilets     :List of 2
#>   .. .. .. ..$ count      : int 1
#>   .. .. .. ..$ category_id: int 179
#>   .. .. ..$ waste_basket:List of 2
#>   .. .. .. ..$ count      : int 7
#>   .. .. .. ..$ category_id: int 180
#>   .. ..$ total_count: int 21
#>   ..$ healthcare      :List of 3
#>   .. ..$ group_id   : int 200
#>   .. ..$ categories :List of 1
#>   .. .. ..$ social_facility:List of 2
#>   .. .. .. ..$ count      : int 1
#>   .. .. .. ..$ category_id: int 210
#>   .. ..$ total_count: int 1
#>   ..$ historic        :List of 3
#>   .. ..$ group_id   : int 220
#>   .. ..$ categories :List of 2
#>   .. .. ..$ memorial:List of 2
#>   .. .. .. ..$ count      : int 1
#>   .. .. .. ..$ category_id: int 237
#>   .. .. ..$ ship    :List of 2
#>   .. .. .. ..$ count      : int 2
#>   .. .. .. ..$ category_id: int 245
#>   .. ..$ total_count: int 3
#>   .. [list output truncated]
#>  $ information:List of 4
#>   ..$ attribution: chr "openrouteservice.org | OpenStreetMap contributors"
#>   ..$ version    : chr "0.1"
#>   ..$ timestamp  : int 1727437218
#>   ..$ query      :List of 2
#>   .. ..$ request : chr "stats"
#>   .. ..$ geometry:List of 2
#>   .. .. ..$ geojson:List of 2
#>   .. .. .. ..$ type       : chr "Point"
#>   .. .. .. ..$ coordinates: num [1:2] 8.8 53.1
#>   .. .. ..$ buffer : int 100